Colmar Academy for Codecademy
This is a very imperfect attempt at the Capstone Project. Due to work I didn't have the amount of time I would have liked to spend on the project, so I injected some um..."creative" design elements. I noticed that many of those are lost when I scale up, so I'll be working on how to do that better in the future. Your comments will be very much appreciated! This is my first time coding anything in this class, and I've really enjoyed it.